chore(oms,pms): move store and test set models to domain packages
