+++ app/oms/stores/models/__init__.py
@@ -1,0 +1,11 @@
+# app/oms/stores/models/__init__.py
+# Domain-owned ORM models for OMS stores.
+
+from app.oms.stores.models.platform_shops import PlatformShop
+from app.oms.stores.models.store import Store, StoreWarehouse
+
+__all__ = [
+    "PlatformShop",
+    "Store",
+    "StoreWarehouse",
+]
